Correct Style (and apply fixes from PhpStan)
