+++ .gitattributes
@@ -6,11 +6,11 @@ core.autocrlf=lf
 .editorconfig
 .gitattributes export-ignore
 .gitignore export-ignore
-.php_cs export-ignore
+.php_cs.dist export-ignore
 .scrutinizer.yml export-ignore
-.styleci.yml export-ignore
 .travis.yml export-ignore
 .github export-ignore
 phpunit.xml.dist export-ignore
+phpstan.neon export-ignore
 /tests export-ignore
 /doc export-ignore
