--- .styleci.yml
@@ -1,25 +1,0 @@
-preset: symfony
-
-risky: true
-linting: true
-
-enabled:
-    - newline_after_open_tag
-    #- ordered_class_elements
-    - ordered_use
-    - phpdoc_order
-    - short_array_syntax
-
-    # Strict
-    - strict_param
-    - declare_strict_types
-
-disabled:
-    - phpdoc_no_empty_return
-
-finder:
-    path:
-        - 'src/'
-        - 'tests/'
-    not-path:
-        - 'Fixtures'
