+++ tests/CachedNativeQueryConditionGeneratorTest.php
@@ -45,7 +45,7 @@ final class CachedNativeQueryConditionGeneratorTest extends OrmTestCase
      */
     protected $conditionGenerator;
 
-    const CACHE_KEY = '8c6844fc8d25be938498e078d917a8fa2e39e00c8fc91834fbd4cc557d99454c';
+    public const CACHE_KEY = '8c6844fc8d25be938498e078d917a8fa2e39e00c8fc91834fbd4cc557d99454c';
 
     public function testGetWhereClauseNoCache()
     {
